Fix ChatMessageRole to be singletons, pull ChatMessage to seperate file
