+++ OpenAI_API/Chat/ChatMessageRole.cs
@@ -12,7 +12,7 @@ namespace OpenAI_API.Chat
 	public class ChatMessageRole : IEquatable<ChatMessageRole>
 	{
 		/// <summary>
-		/// Contructor is prvate to force usage of strongly typed values
+		/// Contructor is private to force usage of strongly typed values
 		/// </summary>
 		/// <param name="value"></param>
 		private ChatMessageRole(string value) { Value = value; }
@@ -37,20 +37,20 @@ namespace OpenAI_API.Chat
 			}
 		}
 
-		private string Value { get; set; }
+		private string Value { get; }
 
 		/// <summary>
 		/// The system message helps set the behavior of the assistant. 
 		/// </summary>
-		public static ChatMessageRole System { get { return new ChatMessageRole("system"); } }
+		public static ChatMessageRole System { get; } = new ChatMessageRole("system");
 		/// <summary>
 		/// The user messages help instruct the assistant. They can be generated by the end users of an application, or set by a developer as an instruction.
 		/// </summary>
-		public static ChatMessageRole User { get { return new ChatMessageRole("user"); } }
+		public static ChatMessageRole User { get; } = new ChatMessageRole("user");
 		/// <summary>
 		/// The assistant messages help store prior responses. They can also be written by a developer to help give examples of desired behavior.
 		/// </summary>
-		public static ChatMessageRole Assistant { get { return new ChatMessageRole("assistant"); } }
+		public static ChatMessageRole Assistant { get; } = new ChatMessageRole("assistant");
 
 		/// <summary>
 		/// Gets the string value for this role to pass to the API
